Renamed folder Handlers to CommandClasses and adjusted namespaces.
