+++ MigFiles/SupportLibraries/ZWaveLib/CommandClasses/CommandClassFactory.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 
-namespace ZWaveLib.Handlers
+namespace ZWaveLib.CommandClasses
 {
     class CommandClassFactory
     {
