+++ MigFiles/SupportLibraries/ZWaveLib/CommandClasses/ThermostatFanMode.cs
@@ -1,4 +1,4 @@
-﻿namespace ZWaveLib.Handlers
+﻿namespace ZWaveLib.CommandClasses
 {
     public enum FanMode
     {
