+++ MigFiles/SupportLibraries/ZWaveLib/CommandClasses/ThermostatHeating.cs
@@ -1,4 +1,4 @@
-﻿namespace ZWaveLib.Handlers
+﻿namespace ZWaveLib.CommandClasses
 {
     public class ThermostatHeating : ICommandClass
     {
