+++ MigFiles/SupportLibraries/ZWaveLib/CommandClasses/ThermostatSetBack.cs
@@ -1,4 +1,4 @@
-﻿namespace ZWaveLib.Handlers
+﻿namespace ZWaveLib.CommandClasses
 {
     public class ThermostatSetBack : ICommandClass
     {
