+++ MigFiles/SupportLibraries/ZWaveLib/CommandClasses/UserCode.cs
@@ -25,7 +25,7 @@ using System;
 using ZWaveLib.Values;
 using System.Collections.Generic;
 
-namespace ZWaveLib.Handlers
+namespace ZWaveLib.CommandClasses
 {
     public class UserCode : ICommandClass
     {
