+++ MigFiles/SupportLibraries/ZWaveLib/ZWaveLib.csproj
@@ -37,33 +37,10 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Handlers\CommandClassFactory.cs" />
-    <Compile Include="Handlers\ThermostatFanMode.cs" />
-    <Compile Include="Handlers\ThermostatFanState.cs" />
-    <Compile Include="Handlers\ThermostatHeating.cs" />
-    <Compile Include="Handlers\ThermostatMode.cs" />
-    <Compile Include="Handlers\ThermostatOperatingState.cs" />
-    <Compile Include="Handlers\ThermostatSetBack.cs" />
-    <Compile Include="Handlers\ThermostatSetPoint.cs" />
     <Compile Include="ICommandClass.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ZWaveMessage.cs" />
     <Compile Include="ZWavePort.cs" />
-    <Compile Include="Handlers\Basic.cs" />
-    <Compile Include="Handlers\SwitchBinary.cs" />
-    <Compile Include="Handlers\SwitchMultilevel.cs" />
-    <Compile Include="Handlers\Alarm.cs" />
-    <Compile Include="Handlers\SensorBinary.cs" />
-    <Compile Include="Handlers\SensorAlarm.cs" />
-    <Compile Include="Handlers\SensorMultilevel.cs" />
-    <Compile Include="Handlers\SceneActivation.cs" />
-    <Compile Include="Handlers\Meter.cs" />
-    <Compile Include="Handlers\Configuration.cs" />
-    <Compile Include="Handlers\Association.cs" />
-    <Compile Include="Handlers\WakeUp.cs" />
-    <Compile Include="Handlers\Battery.cs" />
-    <Compile Include="Handlers\MultiInstance.cs" />
-    <Compile Include="Handlers\ManufacturerSpecific.cs" />
     <Compile Include="Utility.cs" />
     <Compile Include="Controller.cs" />
     <Compile Include="ZWaveNode.cs" />
@@ -71,14 +48,37 @@
     <Compile Include="Values\EnergyValue.cs" />
     <Compile Include="Values\SensorValue.cs" />
     <Compile Include="Values\UserCodeValue.cs" />
-    <Compile Include="Handlers\UserCode.cs" />
     <Compile Include="Values\ZWaveValue.cs" />
     <Compile Include="ZWaveEvent.cs" />
-    <Compile Include="Handlers\Crc16Encapsulated.cs" />
     <Compile Include="Enums\GenericType.cs" />
     <Compile Include="Enums\Command.cs" />
     <Compile Include="Enums\CommandClass.cs" />
     <Compile Include="Enums\Function.cs" />
+    <Compile Include="CommandClasses\CommandClassFactory.cs" />
+    <Compile Include="CommandClasses\ThermostatFanMode.cs" />
+    <Compile Include="CommandClasses\ThermostatFanState.cs" />
+    <Compile Include="CommandClasses\ThermostatHeating.cs" />
+    <Compile Include="CommandClasses\ThermostatMode.cs" />
+    <Compile Include="CommandClasses\ThermostatOperatingState.cs" />
+    <Compile Include="CommandClasses\ThermostatSetBack.cs" />
+    <Compile Include="CommandClasses\ThermostatSetPoint.cs" />
+    <Compile Include="CommandClasses\Basic.cs" />
+    <Compile Include="CommandClasses\SwitchBinary.cs" />
+    <Compile Include="CommandClasses\SwitchMultilevel.cs" />
+    <Compile Include="CommandClasses\Alarm.cs" />
+    <Compile Include="CommandClasses\SensorBinary.cs" />
+    <Compile Include="CommandClasses\SensorAlarm.cs" />
+    <Compile Include="CommandClasses\SensorMultilevel.cs" />
+    <Compile Include="CommandClasses\SceneActivation.cs" />
+    <Compile Include="CommandClasses\Meter.cs" />
+    <Compile Include="CommandClasses\Configuration.cs" />
+    <Compile Include="CommandClasses\Association.cs" />
+    <Compile Include="CommandClasses\WakeUp.cs" />
+    <Compile Include="CommandClasses\Battery.cs" />
+    <Compile Include="CommandClasses\MultiInstance.cs" />
+    <Compile Include="CommandClasses\ManufacturerSpecific.cs" />
+    <Compile Include="CommandClasses\UserCode.cs" />
+    <Compile Include="CommandClasses\Crc16Encapsulated.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
